stats: refine mutex thread safety annotations
Update the mutex thread safety annotations in ThreadLocalStoreImpl to be
compatible with clang 18.

The lock usage here is unusual in that ScopeImpl uses a mutex from its
parent ThreadLocalStoreImpl. There is an existing
ABSL_ASSERT_EXCLUSIVE_LOCK annotation on centralCacheLockHeld() but it
is ineffective because it applies only _after_ the method returns.

While the thread safety analysis does not support aliasing, we can help
it out by adding a no-op method that is annotated so as to convert a
ThreadLocalStoreImpl lock requirement into an assertion about the
ScopeImpl parent lock reference. This is sound if the ScopeImpl's parent
reference is in fact to the same ThreadLocalStoreImpl.

This lets us use a normal ABSL_EXCLUSIVE_LOCKS_REQUIRED annotation on
centralCacheLockHeld().

Signed-off-by: Kenneth Jenkins <51246568+kenjenkins@users.noreply.github.com>